Refactor main page
