+++ web/src/features/main/models/Repository.ts
@@ -1,0 +1,11 @@
+export type Repository = {
+    name: string,
+    accessMode: RepositoryAccessMode,
+    updatedAt: string,
+    description?: string,
+}
+
+export enum RepositoryAccessMode {
+    PUBLIC = "Public",
+    PRIVATE = "Private",
+}
